requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
pydantic==2.3.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
pandas
pyyaml
pytest
ruff
black
pyspark
requests
python-dotenv
pytest-asyncio
pytest-cov
fakeredis
testcontainers
httpx
httpx[http2]
httpx[http2,cli]
httpx[http2,cli,keepalive]
flake8